% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_property.R
\name{pep_property}
\alias{pep_property}
\title{peptide property calculation}
\usage{
pep_property(pep)
}
\arguments{
\item{pep}{The input peptide sequence}
}
\value{
The output from \code{\link{pep_property}}
}
\description{
calculate various different peptide properties according to the peptide sequence.
}
\examples{
pep_property("AAAAAAAAAA")
}
