% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_iso_conc.R
\name{predict_iso_conc}
\alias{predict_iso_conc}
\title{Predict isopropanol elution concentration}
\usage{
predict_iso_conc(pep_property)
}
\arguments{
\item{pep_property}{The input peptide property calculated by function pep_property()}
}
\value{
The output from \code{\link{predict_iso_conc}}
}
\description{
Predict isopropanol elution concentration for a given peptide sequence
}
\examples{
seq_property<-pep_property("AAAAAAAAAA")
predict_iso_conc(seq_property)
}
